requirements.txt
neo4j
langchain>=0.3.0
langchain-community
langchain-openai
langchain-text-splitters
python-dotenv